Finished refactoring rooms
